+++ src/main/java/solutions/fairdata/openrefine/metadata/dto/audit/EventSource.java
@@ -20,24 +20,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.openrefine.metadata.commands.response.metadata;
+package solutions.fairdata.openrefine.metadata.dto.audit;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class MetadataSpecResponse {
-
-    public String status;
-    public String spec;
-
-    public MetadataSpecResponse(String spec) {
-        this.status = "ok";
-        this.spec = spec;
-    }
+public enum EventSource {
+    FDP_CONNECTION,
+    FDP_METADATA,
+    SERVICE,
+    STORAGE,
+    AUDIT,
+    TYPEHINTS,
+    FRONTEND,
+    SETTINGS;
 }