+++ src/main/resources/module/scripts/dialogs/metadata-form-dialog.js
@@ -1,7 +1,7 @@
 /* global DOM, DialogSystem, Refine, MetadataApiClient, StoreDataDialog */
 
 class MetadataFormDialog {
-    constructor(specs, callbackFn, prefill) {
+    constructor(apiClient, specs, callbackFn, prefill) {
         this.frame = $(DOM.loadHTML("metadata", "scripts/dialogs/metadata-form-dialog.html"));
         this.elements = DOM.bind(this.frame);
         this.level = null;
@@ -9,15 +9,15 @@ class MetadataFormDialog {
         this.type = specs.id;
         this.specs = specs;
         this.callbackFn = callbackFn;
-        this.apiClient = new MetadataApiClient();
+        this.apiClient = apiClient;
+        this.shaclSpec = "";
 
         this.datalists = new Set();
 
-        const prefillMap = prefill || new Map();
+        this.prefill = prefill || new Map();
 
         this.initBasicTexts();
-        this.createForm();
-        this.fillForm(prefillMap);
+        this.getSpec();
         this.bindActions();
     }
 
@@ -73,6 +73,15 @@ class MetadataFormDialog {
             }
             input.trigger("change");
         }
+    }
+
+    getSpec() {
+        this.apiClient.getMetadataSpec(this.type, [
+            (result) => {
+                this.shaclSpec = result.spec;
+                this.createForm();
+            }
+        ], []);
     }
 
     createForm() {
@@ -133,6 +142,8 @@ class MetadataFormDialog {
                 formGroupId
             );
         }
+
+        this.fillForm(this.prefill);
     }
 
     displayError(errorName, errorMessage) {
@@ -429,8 +440,8 @@ class MetadataFormDialog {
     }
 
     // launcher
-    static createAndLaunch(specs, callbackFn, prefill) {
-        const dialog = new MetadataFormDialog(specs, callbackFn, prefill);
+    static createAndLaunch(apiClient, specs, callbackFn, prefill) {
+        const dialog = new MetadataFormDialog(apiClient, specs, callbackFn, prefill);
         dialog.launch();
     }
 }