+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0]
+
+### Added
+
+- Audit log (per-project) with possibility to browse, filter, and clear using 
+  a simple dialog in client-side
+
+### Changed
+
+- Prepared getting and passing SHACL specs from FDP to OpenRefine frontend
+- Updated for compatibility with newer FAIR Data Point 1.3.0
+
+### Fixed
+
+- Persisting information about stored metadata
+
 ## [1.2.0]
 
 ### Added
@@ -64,8 +80,8 @@ Initial version based on reproducing functionality of deprecated [FAIRifier](htt
 - About dialog with basic information about the extension and its compatibility
 - Report a bug link to create a GitHub issue easily
 
-[Unreleased]: /../../compare/v1.1.0...develop
+[Unreleased]: /../../compare/v1.3.0...develop
 [1.0.0]: /../../tree/v1.0.0
 [1.1.0]: /../../tree/v1.1.0
 [1.2.0]: /../../tree/v1.2.0
-
+[1.3.0]: /../../tree/v1.3.0